replay viewer: add 1/2x playback speed; Space already pauses the shipped page
The replay transport gains a 0.5x speed: command '5' selects a replay-only
half speed (ReplayHalfSpeedIndex) whose advance spends one tick every other
presentation frame (halfPhase parity), '-' floors on it and '+' climbs back
out, the chrome grows a 0.5x chip, and the frame JSON's sp field carries 0.5
while it is active (ViewChrome.speed is a float now). The wire constants
prepend 0.5 to the emitted speeds so the engine list drives the chips; the
global stays CTF_WIRE because chrome_common.js reads that name and the
replay-viewer Dockerfile pins the line. The one shipped page (index.html
from client/replay_broadcast.html) already binds Space to play/pause and
already forwards digit keys, so '5' works from the keyboard unchanged.

+++ src/hidden_agenda/broadcast.nim
@@ -34,7 +34,9 @@ type
     maxTick*, maxTicks*, depositTarget*: int
     phase*: string                ## lobby | playing | gameover
     playing*, looping*, skipping*, fastForward*, enabled*: bool
-    speed*: int
+    speed*: float
+      ## What the speed chips highlight: 0.5 at the replay-only 1/2x, else
+      ## the integer playback speed.
     activeCrew*, impostorSlot*, freezeCooldown*, freezeCooldownTicks*: int
     meetingNumber*, meetingIn*: int
     meetingCause*: string
